nestor_slack_writer/requirements.txt
slackclient==2.9.1
pika==1.1.0